Add workaround documentation for numeric network identities
cli4clj wraps clojure.main/repl. The clojure `read` function tries to parse numeric
network identies as numbers, and fails:

> 12345/abc
Invalid number

The read function isn't pluggable in cli4clj, so I've filed a bug.
https://github.com/ruedigergad/cli4clj/issues/4

In the meantime, adding documentation of the workaround, which prevents the reader from
trying to parse the input as a number.

FC-1297